# doc/api.yaml
openapi: 3.0.0
paths:
  /:
    get:
      operationId: AppController_getHello
      parameters: []
      responses:
        '200':
          description: ''
  /user:
    get:
      operationId: UserController_findAll
      parameters: []
      responses:
        '200':
          description: ''
    post:
      operationId: UserController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserDto'
      responses:
        '201':
          description: ''
  /user/{id}:
    get:
      operationId: UserController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
    put:
      operationId: UserController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePasswordDto'
      responses:
        '200':
          description: ''
    delete:
      operationId: UserController_delete
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: ''
  /artist:
    get:
      operationId: ArtistsController_findAll
      parameters: []
      responses:
        '200':
          description: ''
    post:
      operationId: ArtistsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateArtistDto'
      responses:
        '201':
          description: ''
  /artist/{id}:
    get:
      operationId: ArtistsController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
    put:
      operationId: ArtistsController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateArtistDto'
      responses:
        '200':
          description: ''
    delete:
      operationId: ArtistsController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: ''
  /favs:
    get:
      operationId: FavoritesController_findAll
      summary: Get all favorites
      parameters: []
      responses:
        '200':
          description: Successfully retrieved all favorites
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Favorites'
      tags: &ref_0
        - Favorites
  /favs/track/{id}:
    post:
      operationId: FavoritesController_addTrack
      summary: Add track to favorites
      parameters:
        - name: id
          required: true
          in: path
          description: Track UUID
          schema:
            example: 123e4567-e89b-12d3-a456-426614174000
            type: string
      responses:
        '201':
          description: Track successfully added to favorites
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Track added to favorites
        '400':
          description: Invalid UUID format
        '422':
          description: Track does not exist
      tags: *ref_0
    delete:
      operationId: FavoritesController_removeTrack
      summary: Remove track from favorites
      parameters:
        - name: id
          required: true
          in: path
          description: Track UUID
          schema:
            example: 123e4567-e89b-12d3-a456-426614174000
            type: string
      responses:
        '204':
          description: Track successfully removed from favorites
        '400':
          description: Invalid UUID format
        '404':
          description: Track not found in favorites
      tags: *ref_0
  /favs/album/{id}:
    post:
      operationId: FavoritesController_addAlbum
      summary: Add album to favorites
      parameters:
        - name: id
          required: true
          in: path
          description: Album UUID
          schema:
            example: 123e4567-e89b-12d3-a456-426614174000
            type: string
      responses:
        '201':
          description: Album successfully added to favorites
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Album added to favorites
        '400':
          description: Invalid UUID format
        '422':
          description: Album does not exist
      tags: *ref_0
    delete:
      operationId: FavoritesController_removeAlbum
      summary: Remove album from favorites
      parameters:
        - name: id
          required: true
          in: path
          description: Album UUID
          schema:
            example: 123e4567-e89b-12d3-a456-426614174000
            type: string
      responses:
        '204':
          description: Album successfully removed from favorites
        '400':
          description: Invalid UUID format
        '404':
          description: Album not found in favorites
      tags: *ref_0
  /favs/artist/{id}:
    post:
      operationId: FavoritesController_addArtist
      summary: Add artist to favorites
      parameters:
        - name: id
          required: true
          in: path
          description: Artist UUID
          schema:
            example: 123e4567-e89b-12d3-a456-426614174000
            type: string
      responses:
        '201':
          description: Artist successfully added to favorites
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Artist added to favorites
        '400':
          description: Invalid UUID format
        '422':
          description: Artist does not exist
      tags: *ref_0
    delete:
      operationId: FavoritesController_removeArtist
      summary: Remove artist from favorites
      parameters:
        - name: id
          required: true
          in: path
          description: Artist UUID
          schema:
            example: 123e4567-e89b-12d3-a456-426614174000
            type: string
      responses:
        '204':
          description: Artist successfully removed from favorites
        '400':
          description: Invalid UUID format
        '404':
          description: Artist not found in favorites
      tags: *ref_0
  /album:
    post:
      operationId: AlbumsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAlbumDto'
      responses:
        '201':
          description: ''
    get:
      operationId: AlbumsController_findAll
      parameters: []
      responses:
        '200':
          description: ''
  /album/{id}:
    get:
      operationId: AlbumsController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
    put:
      operationId: AlbumsController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAlbumDto'
      responses:
        '200':
          description: ''
    delete:
      operationId: AlbumsController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: ''
  /track:
    get:
      operationId: TracksController_findAll
      parameters: []
      responses:
        '200':
          description: ''
    post:
      operationId: TracksController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTrackDto'
      responses:
        '201':
          description: ''
  /track/{id}:
    get:
      operationId: TracksController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
    put:
      operationId: TracksController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTrackDto'
      responses:
        '200':
          description: ''
    delete:
      operationId: TracksController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: ''
info:
  title: API Documentation
  description: The API description
  version: '1.0'
  contact: {}
tags: []
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    CreateUserDto:
      type: object
      properties: {}
    UpdatePasswordDto:
      type: object
      properties: {}
    CreateArtistDto:
      type: object
      properties: {}
    UpdateArtistDto:
      type: object
      properties: {}
    Favorites:
      type: object
      properties:
        artists:
          description: List of favorite artist UUIDs
          example:
            - 123e4567-e89b-12d3-a456-426614174000
            - 223e4567-e89b-12d3-a456-426614174001
          default: []
          type: array
          items:
            type: string
        albums:
          description: List of favorite album UUIDs
          example:
            - 323e4567-e89b-12d3-a456-426614174002
            - 423e4567-e89b-12d3-a456-426614174003
          default: []
          type: array
          items:
            type: string
        tracks:
          description: List of favorite track UUIDs
          example:
            - 523e4567-e89b-12d3-a456-426614174004
            - 623e4567-e89b-12d3-a456-426614174005
          default: []
          type: array
          items:
            type: string
      required:
        - artists
        - albums
        - tracks
    CreateAlbumDto:
      type: object
      properties: {}
    UpdateAlbumDto:
      type: object
      properties: {}
    CreateTrackDto:
      type: object
      properties: {}
    UpdateTrackDto:
      type: object
      properties: {}